refactor(ui): вынести четыре экрана из DocumentSetsPage по файлам (#488)
Файл был не страницей, а картой маршрутов и четырьмя независимыми экранами:
SetDetail 444 строки, SectionDetail 169, ConstructionDetail 162,
ConstructionsList 153, и 74 обращения к данным на весь файл.

Резал по шву, а не по счётчику строк. Экраны не разделяют между собой ничего,
кроме импортов: у каждого свои параметры маршрута, свои хуки, свой
ListDetailShell и свой локальный тип панели; наружу экспортировалось единственное
имя — сама страница. Поэтому это перенос, а не перестройка.

DocumentSetsPage.tsx 1051 → 22 строки, остальное по файлам рядом.
DocumentSearchPanel уехал вместе с ConstructionsList — им пользуется только она.
Импорты разнесены по фактическому использованию, локальные типы панелей — к
своим экранам.

Из четырёх крупных страниц выбрана эта: у editor/index.tsx шов другого качества —
там половина файла один компонент RequisitesTab на 604 строки, и его разделение
потребовало бы настоящей декомпозиции с другим риском.

Проверено живьём по всем шести маршрутам (список строек, стройка,
стройка/наборы, раздел, комплект, комплект/сверка): содержимое отрисовано,
ошибок страницы нет. Тесты у нас на чистых функциях, регрессию отрисовки они бы
не поймали. tsc -b чист, 212 тестов зелёные.
